+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## x.x.x
+- Fixed sending '--' as carrier name value due to deprecated calls from iOS version 16.4
+
 ## 23.8.3
 - Added `requestDropAgeHours` initial config property to set a time limit after which the requests would be removed if not sent to the server
 - Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'