+++ CHANGELOG.md
@@ -1,11 +1,9 @@
-## x.x.x
-- Fixed sending '--' as carrier name value due to deprecated calls from iOS version 16.4
-
 ## 23.8.3
 - Added `requestDropAgeHours` initial config property to set a time limit after which the requests would be removed if not sent to the server
 - Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'
 - Added a call to enroll users to A/B tests when getting all remote config values: 'getAllValuesAndEnroll'
 
+- Fixed sending '--' as carrier name due to platform changes from iOS version 16.4. This version and above will now not send any carrier information due to platform limitations.
 - Mitigated an issue where users could not enroll to an A/B tests if enrollment request has failed
 
 ## 23.8.2